refactor: code gold
